requirements.txt
praw
pandas
sentence-transformers
faiss-cpu
numpy

